tests/requirements.txt
pytest
docker
requests
pytest-html
nuvla-api
gitpython
six